// src/app/shared/model/response/srvy-task/res-srvy-task.model.ts
export class ResSrvyTaskObj {
    SrvyTaskId: number;
    SrvyOrderId: number;
    SrvyTaskNo: string;
    MrSurveyTypeCode: string;
    MrSurveyTaskStatCode: string;
    MrSrvyObjTypeCode: string;
    RefNo:string;
    SurveyorId: number;
    RetrieveDt: Date;
    ResultDt: Date;
    ReviewByRefUserId: number;
    ReviewDt: Date;
    ReviewNotes: string;
    IsAddtReq: boolean;
    PrevSurveyTaskNo: string;
    Result: string;
    CustNo: string;
    MrCustModelCode: string;
    CustName: string;
    CustAddr:string;
    CustPhone: string;
    Zipcode: string;
    AssignDt: Date;
    SrvyFormSchmId: number;
    Notes: string;
    RefReasonId: number;
    PrevSurveyorId: number;
    MobileAssignmentId: number;
    AreaCode1: string;
    AreaCode2: string;
    AreaCode3: string;
    AreaCode4: string;
    City: string;
    Addr: string;
    SrvySubjectId: number;
    SrvyObjectId: number;
    RowVersion: string;

    constructor() {
        this.SrvyTaskId = 0;
        this.SrvyOrderId = 0;
        this.SrvyTaskNo = "";
        this.MrSurveyTypeCode = "";
        this.MrSurveyTaskStatCode = "";
        this.MrSrvyObjTypeCode = "";
        this.RefNo = "";
        this.SurveyorId = 0;
        this.RetrieveDt = new Date();
        this.ResultDt = new Date();
        this.ReviewByRefUserId = 0;
        this.ReviewDt = new Date();
        this.ReviewNotes = "";
        this.IsAddtReq = false;
        this.PrevSurveyTaskNo = "";
        this.Result = "";
        this.CustNo = "";
        this.MrCustModelCode = "";
        this.CustName = "";
        this.CustAddr = "";
        this.CustPhone = "";
        this.Zipcode = "";
        this.AssignDt
        this.SrvyFormSchmId = 0;
        this.Notes = "";
        this.RefReasonId = 0;
        this.PrevSurveyorId = 0;
        this.MobileAssignmentId = 0;
        this.AreaCode1 = "";
        this.AreaCode2 = "";
        this.AreaCode3 = "";
        this.AreaCode4 = "";
        this.City = "";
        this.Addr = "";
        this.SrvySubjectId = 0;
        this.SrvyObjectId = 0;
        this.RowVersion = "";
    }
}  