// src/app/shared/model/verf-result-obj.model.ts
export class VerfResultObj {
    VerfResultId: number;
    VerfResultNo: string;
    TrxRefNo: string;
    VerfDt: Date;
    EmpNo: string;
    MrVerfResultStatCode: string;
    MrVerfTrxTypeCode: string;
    LobCode: string;
    LobName: string;
    Notes: string;
    RowVersion : string;
    constructor() { this.VerfResultId = 0; }
}  